moved main menu scene to planet selection scene
