+++ Assets/Scripts/PlanetSelection/PlanetSelectionUI/PlanetSelectionUIManager.cs
@@ -5,17 +5,17 @@ public class PlanetSelectionUIManager : MonoBehaviour
     public Canvas planetUICanvas;
     public Canvas planetSelectionCanvas;
     public Canvas levelUICanvas;
+    public Canvas mainMenuCanvas;
     public LevelUI levelUI;
 
     void Start()
     {
-        levelUICanvas.gameObject.SetActive(false);
+        mainMenuCanvas.gameObject.SetActive(true);
 
-        if (!PlayerPrefs.HasKey("LastPlanet"))
-        {
-            planetSelectionCanvas.gameObject.SetActive(true);
-            planetUICanvas.gameObject.SetActive(false);
-        }
+        levelUICanvas.gameObject.SetActive(false);
+        planetSelectionCanvas.gameObject.SetActive(false);
+        planetUICanvas.gameObject.SetActive(false);
+
     }
 
     public void SetPlanetUICanvasActive(bool active)
@@ -33,7 +33,12 @@ public class PlanetSelectionUIManager : MonoBehaviour
         levelUICanvas.gameObject.SetActive(active);
     }
 
-    public void OpenLevelUI(int LevelID) 
+    public void SetMainMenuCanvasActive(bool active)
+    {
+        mainMenuCanvas.gameObject.SetActive(active);
+    }
+
+    public void OpenLevelUI(int LevelID)
     {
         levelUI.DisplayLevelInfo(LevelID);
         levelUICanvas.gameObject.SetActive(true);
