+++ ProjectSettings/EditorBuildSettings.asset
@@ -5,7 +5,7 @@ EditorBuildSettings:
   m_ObjectHideFlags: 0
   serializedVersion: 2
   m_Scenes:
-  - enabled: 1
+  - enabled: 0
     path: Assets/Scenes/MainMenuScene.unity
     guid: f2960a81afcae7447b538f2255d2633c
   - enabled: 0
